pyproject.toml
[project]
name = "glitchnodes"
description = "GlitchNodes is a collection of image processing nodes designed for ComfyUI that specializes in creating glitch art and retro effects."
version = "1.0.6"
license = {file = "LICENSE"}
dependencies = ["numpy", "torch", "Pillow", "opencv-python", "scipy"]

[project.urls]
Repository = "https://github.com/pxl-pshr/GlitchNodes"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "pxlpshr"
DisplayName = "GlitchNodes"
Icon = ""
